Crate chain_connector

Structs

CCInitParams

ChainConnector

CurrentEpochFunction

@dev Returns current epoch @return current epoch number function currentEpoch() external view returns (uint256);

DifficultyFunction

function difficulty() external view returns (bytes32);

EpochDurationFunction

@dev Returns epoch duration @return epochDuration in seconds function epochDuration() external view returns (uint256);

GetCommitmentFunction

@dev Returns the commitment info @param commitmentId Commitment id @return info commitment info function getCommitment(bytes32 commitmentId) external view returns (CommitmentView memory);

GetCommitmentStatusFunction

@dev Returns the commitment status @param commitmentId Commitment id @return status commitment status function getStatus(bytes32 commitmentId) external view returns (CCStatus);

GetComputePeerFunction

struct ComputePeer { bytes32 offerId; bytes32 commitmentId; uint256 unitCount; address owner; } @dev Returns the compute peer info function getComputePeer(bytes32 peerId) external view returns (ComputePeer memory);

GetComputeUnitsFunction

@dev Returns the compute units info of a peer function getComputeUnits(bytes32 peerId) external view returns (ComputeUnitView[] memory);

GetDealStatusFunction

function getStatus() public view returns (Status)

GetGlobalNonceFunction

function getGlobalNonce() external view returns (bytes32);

InitTimestampFunction

@dev Returns epoch init timestamp @return initTimestamp in seconds function initTimestamp() external view returns (uint256);

ReturnComputeUnitFromDeal

@dev Return the compute unit from a deal function returnComputeUnitFromDeal(bytes32 unitId) external;

SubmitProofFunction

@dev Submits a proof for the commitment @param unitId Compute unit id which provied the proof @param globalUnitNonce The global nonce of the unit for calculating the target hash @param localUnitNonce The local nonce of the unit for calculating the target hash. It’s the proof @param targetHash The target hash of this proof function submitProof(bytes32 unitId, bytes32 globalUnitNonce, bytes32 localUnitNonce, bytes32 targetHash) external;

Enums

ConnectorError