Module chain_connector::Capacity

Module containing a contract’s types and functions.

contract Capacity {
    error TooManyProofs();
    error CapacityCommitmentIsNotActive(CCStatus status);
    function getGlobalNonce() external view returns (bytes32);
    function getStatus(bytes32 commitmentId) external view returns (CCStatus);
    function submitProof(bytes32 unitId, bytes32 localUnitNonce, bytes32 resultHash) external;
}

Structs

CapacityCommitmentIsNotActive

@dev Capacity commitment is not active Custom error with signature CapacityCommitmentIsNotActive(uint8) and selector 0x0852c720.

TooManyProofs

@dev Throws if peer sent too many proofs for the commitment by unit per epoch Custom error with signature TooManyProofs() and selector 0xe5d50da7.

getGlobalNonceCall

Function with signature getGlobalNonce() and selector 0x92163b4b.

getGlobalNonceReturn

Container type for the return parameters of the getGlobalNonce() function.

getStatusCall

@dev Returns the commitment status @param commitmentId Commitment id @return status commitment status Function with signature getStatus(bytes32) and selector 0x5de28ae0.

getStatusReturn

@dev Returns the commitment status @param commitmentId Commitment id @return status commitment status Container type for the return parameters of the getStatus(bytes32) function.

submitProofCall

@dev Submits a proof for the commitment @param unitId Compute unit id which provied the proof @param localUnitNonce The local nonce of the unit for calculating the target hash. It’s the proof @param resultHash The target hash of this proof Function with signature submitProof(bytes32,bytes32,bytes32) and selector 0x4ece5685.

submitProofReturn

@dev Submits a proof for the commitment @param unitId Compute unit id which provied the proof @param localUnitNonce The local nonce of the unit for calculating the target hash. It’s the proof @param resultHash The target hash of this proof Container type for the return parameters of the submitProof(bytes32,bytes32,bytes32) function.

Enums

CapacityCalls

Container for all the Capacity function calls.

CapacityErrors

Container for all the Capacity custom errors.