Module chain_connector::Core

Module containing a contract’s types and functions.

contract Core {
    function capacity() external view returns (address);
    function market() external view returns (address);
    function currentEpoch() external view returns (uint256);
    function epochDuration() external view returns (uint256);
    function initTimestamp() external view returns (uint256);
    function difficulty() external view returns (bytes32);
    function minProofsPerEpoch() external view returns (uint256);
    function maxProofsPerEpoch() external view returns (uint256);
}

Structs

capacityCall

Function with signature capacity() and selector 0x5cfc1a51.

capacityReturn

Container type for the return parameters of the capacity() function.

currentEpochCall

@dev Returns current epoch @return current epoch number Function with signature currentEpoch() and selector 0x76671808.

currentEpochReturn

@dev Returns current epoch @return current epoch number Container type for the return parameters of the currentEpoch() function.

difficultyCall

@dev Returns the difficulty for CCP Function with signature difficulty() and selector 0x19cae462.

difficultyReturn

@dev Returns the difficulty for CCP Container type for the return parameters of the difficulty() function.

epochDurationCall

@dev Returns epoch duration @return epochDuration in seconds Function with signature epochDuration() and selector 0x4ff0876a.

epochDurationReturn

@dev Returns epoch duration @return epochDuration in seconds Container type for the return parameters of the epochDuration() function.

initTimestampCall

@dev Returns epoch init timestamp @return initTimestamp in seconds Function with signature initTimestamp() and selector 0x7358c57a.

initTimestampReturn

@dev Returns epoch init timestamp @return initTimestamp in seconds Container type for the return parameters of the initTimestamp() function.

marketCall

Function with signature market() and selector 0x80f55605.

marketReturn

Container type for the return parameters of the market() function.

maxProofsPerEpochCall

@dev Returns the max randomX proofs per epoch Function with signature maxProofsPerEpoch() and selector 0x280c5e51.

maxProofsPerEpochReturn

@dev Returns the max randomX proofs per epoch Container type for the return parameters of the maxProofsPerEpoch() function.

minProofsPerEpochCall

@dev Returns the min required randomX proofs per epoch for the 1 CU. @dev If lower than this - CU is failed and CC slashed. Function with signature minProofsPerEpoch() and selector 0xe79c6270.

minProofsPerEpochReturn

@dev Returns the min required randomX proofs per epoch for the 1 CU. @dev If lower than this - CU is failed and CC slashed. Container type for the return parameters of the minProofsPerEpoch() function.

Enums

CoreCalls

Container for all the Core function calls.