Module chain_connector::Deal

Module containing a contract’s types and functions.

contract Deal {
    struct CIDV1 { bytes4 prefixes; bytes32 hash; }
    enum Status { INSUFFICIENT_FUNDS, ACTIVE, ENDED, NOT_ENOUGH_WORKERS, SMALL_BALANCE }
    struct Worker { bytes32 offchainId; bytes32 onchainId; bytes32 peerId; address provider; uint256 joinedEpoch; bytes32[] computeUnitIds; }
    function getStatus() external view returns (Status);
    function appCID() external view returns (CIDV1 memory);
    function activateWorker(bytes32 onchainId, bytes32 offchainId);
    function removeWorker(bytes32 onchainId) external;
    function getWorkers() external view returns (Worker[] memory);
}

Structs

CIDV1

Worker

activateWorkerCall

@dev Set offchain worker ID for a corresponding onchain worker for a deal Function with signature activateWorker(bytes32,bytes32) and selector 0x22ebd58a.

activateWorkerReturn

@dev Set offchain worker ID for a corresponding onchain worker for a deal Container type for the return parameters of the activateWorker(bytes32,bytes32) function.

appCIDCall

@dev Returns the app CID Function with signature appCID() and selector 0x9bc66868.

appCIDReturn

@dev Returns the app CID Container type for the return parameters of the appCID() function.

getStatusCall

@dev Returns the status of the deal Function with signature getStatus() and selector 0x4e69d560.

getStatusReturn

@dev Returns the status of the deal Container type for the return parameters of the getStatus() function.

getWorkersCall

@dev Returns workers Function with signature getWorkers() and selector 0xc1499f71.

getWorkersReturn

@dev Returns workers Container type for the return parameters of the getWorkers() function.

removeWorkerCall

@dev Removes worker from the deal Function with signature removeWorker(bytes32) and selector 0xbc126f1f.

removeWorkerReturn

@dev Removes worker from the deal Container type for the return parameters of the removeWorker(bytes32) function.

Enums

DealCalls

Container for all the Deal function calls.

Status