Module chain_connector::Offer

Module containing a contract’s types and functions.

contract Offer {
    struct ComputePeer { bytes32 offerId; bytes32 commitmentId; uint256 unitCount; address owner; }
    struct ComputeUnit { bytes32 id; address deal; uint256 startEpoch; }
    function getComputePeer(bytes32 peerId) external view returns (ComputePeer memory);
    function getComputeUnits(bytes32 peerId) external view returns (ComputeUnit[] memory);
    function returnComputeUnitFromDeal(bytes32 unitId) external;
}

Structs

ComputePeer

ComputeUnit

getComputePeerCall

@dev Returns the compute peer info Function with signature getComputePeer(bytes32) and selector 0x86e68254.

getComputePeerReturn

@dev Returns the compute peer info Container type for the return parameters of the getComputePeer(bytes32) function.

getComputeUnitsCall

@dev Returns the compute units info of a peer Function with signature getComputeUnits(bytes32) and selector 0xb6015c6e.

getComputeUnitsReturn

@dev Returns the compute units info of a peer Container type for the return parameters of the getComputeUnits(bytes32) function.

returnComputeUnitFromDealCall

@dev Return the compute unit from a deal Function with signature returnComputeUnitFromDeal(bytes32) and selector 0x14afcd52.

returnComputeUnitFromDealReturn

@dev Return the compute unit from a deal Container type for the return parameters of the returnComputeUnitFromDeal(bytes32) function.

Enums

OfferCalls

Container for all the Offer function calls.